perf(useEffect): reduce re-render count by useEffect(8 -\> 4)
- useEffect는 끊임없이 상태를 감지하고, 해당 상태 변경에 따른 리렌더링을 진행한다. 이는 상태가 잦은 변경이 일어날 경우 비용증가를 야기한다.
- 따라서 useEffect를 합치거나, 쓸모 없는 useEffect는 제거하는 형식으로 전개했다.
- 추가로, layout는 다른 곳에서 재사용할 수 있으므로 발생할 수 있는 side effect를 하위 페이지로 넘겼다.
- 결과적으로 약 8번의 리렌더링이 필연적으로 일어났는데, 4번의 리렌더링까지 줄여내는 데 성공했다.
